+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Released on Wednesday, September 9 2026
 
 - Fixed unresolved converter re-entry (#243) @sebastienros
 - Fixed issue with computation of `transform` functions
+- Added explicit pseudo class handling via `SetPseudoClass`
 
 # 1.1.0